Operators & Values edits
In the operators section, I didn't think the exponentiation example at the top was 'more or less clear' so I removed that phrase. Also, I expanded the explanation of it. Also noted that % for modulo is common in other languages as well.

Fixed numeric typo at top of values chapter and removed note that seemed unnecessary. Also, removed mention of 'classes' as something they will get in a later chapter. Also removed mention of dictionaries and sets, since we don't cover them.